Upload binaries to transfer.sh
We no longer publish a new release after a build because it breaks
other workflows.

Publish the `continuous` release triggers COPR rebuilds and tagging
a version will replace the release on GitHub, meaning that we have
to wait for the build before publishing a release and this slows
down the workflow.