Port build system to CMake
Qbs is now deprecated:
http://blog.qt.io/blog/2018/10/29/deprecation-of-qbs/
http://lists.qt-project.org/pipermail/development/2018-October/034023.html

The only viable alternative to qmake left is CMake.
Chances are that it will be used to build Qt 6, and in that
case QtCreator support will hopefully be improved.

CMake is also widely adopted and thanks to the server mode
it is easier to use with QtCreator nowadays, although some
basic features such as adding files from the IDE are not
supported yet.  Not to mention missing Android integration.

Despite Qbs being a clearly superior build tool than CMake,
we are forced to abandon it in favor of the status quo.

Even though Qbs is still alive and some companies appear
to be willing to at least maintain QtCreator integration
in its current status, we are not guaranteed that the
build system will evolve with all the promised feature.

So thanks to the Qbs team for letting us dream, but we
can't risk Liri health for a build system that will
potentially be stale.

The official build system is now CMake.

Qbs files will probably stay here for a while, but we will
not update them anymore.  If, at some point, Qbs will get
momentum we will evaluate the situation again.

Issue: #29